requirements.txt
streamlit
pandas
openpyxl
pillow
playwright
